+++ continuous_integration/build_test_ext.sh
@@ -1,5 +1,0 @@
-
-cd tests/_openmp
-COVERAGE_PROCESS_START=''
-python setup.py build_ext -i || echo 'No openmp'
-cd ../..